+++ modules/features/search/src/main/java/au/com/shiftyjelly/pocketcasts/search/SearchViewModel.kt
@@ -5,13 +5,12 @@ import androidx.lifecycle.viewModelScope
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTracker
 import au.com.shiftyjelly.pocketcasts.analytics.SourceView
-import au.com.shiftyjelly.pocketcasts.models.entity.BaseEpisode
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.models.to.EpisodeItem
 import au.com.shiftyjelly.pocketcasts.models.to.FolderItem
+import au.com.shiftyjelly.pocketcasts.models.to.ImprovedSearchResultItem
 import au.com.shiftyjelly.pocketcasts.models.to.SearchAutoCompleteItem
 import au.com.shiftyjelly.pocketcasts.models.to.SearchHistoryEntry
-import au.com.shiftyjelly.pocketcasts.repositories.podcast.EpisodeManager
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.PodcastManager
 import au.com.shiftyjelly.pocketcasts.repositories.searchhistory.SearchHistoryManager
 import au.com.shiftyjelly.pocketcasts.search.SearchResultsFragment.Companion.ResultsType
@@ -21,11 +20,8 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.reactive.asFlow
-import kotlinx.coroutines.rx2.rxMaybe
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
 
 @HiltViewModel
@@ -34,7 +30,6 @@ class SearchViewModel @Inject constructor(
     private val searchHistoryManager: SearchHistoryManager,
     private val podcastManager: PodcastManager,
     private val analyticsTracker: AnalyticsTracker,
-    private val episodeManager: EpisodeManager,
 ) : ViewModel() {
     var isFragmentChangingConfigurations: Boolean = false
     var showSearchHistory: Boolean = true
@@ -51,28 +46,26 @@ class SearchViewModel @Inject constructor(
             viewModelScope.launch {
                 searchHandler.searchSuggestions
                     .collect { operation ->
-                        _state.update {
-                            if (it is SearchUiState.Idle || it is SearchUiState.Suggestions) {
-                                when (operation) {
-                                    is SearchUiState.SearchOperation.Error -> {
-                                        analyticsTracker.track(
-                                            AnalyticsEvent.IMPROVED_SEARCH_SUGGESTIONS_FAILED,
-                                            mapOf(
-                                                "source" to "discover",
-                                            ),
-                                        )
-                                    }
-                                    else -> Unit
+                        showSearchHistory = false
+                        _state.update { uiState ->
+                            when (operation) {
+                                is SearchUiState.SearchOperation.Error -> {
+                                    analyticsTracker.track(
+                                        AnalyticsEvent.IMPROVED_SEARCH_SUGGESTIONS_FAILED,
+                                        mapOf(
+                                            "source" to source,
+                                        ),
+                                    )
                                 }
 
-                                // only show loading for the initial query when autocomplete results are empty
-                                if (((it as? SearchUiState.Suggestions)?.operation as? SearchUiState.SearchOperation.Success)?.results?.isNotEmpty() == true && operation is SearchUiState.SearchOperation.Loading) {
-                                    it
-                                } else {
-                                    SearchUiState.Suggestions(operation = operation)
-                                }
+                                else -> Unit
+                            }
+
+                            // only show loading for the initial query when autocomplete results are empty
+                            if (((uiState as? SearchUiState.Suggestions)?.operation as? SearchUiState.SearchOperation.Success)?.results?.isNotEmpty() == true && operation is SearchUiState.SearchOperation.Loading) {
+                                uiState
                             } else {
-                                it
+                                SearchUiState.Suggestions(operation = operation)
                             }
                         }
                     }
@@ -80,14 +73,23 @@ class SearchViewModel @Inject constructor(
         }
 
         viewModelScope.launch {
-            searchHandler.searchResults.collect {
-                if (_state.value is SearchUiState.Results || !FeatureFlag.isEnabled(Feature.IMPROVED_SEARCH_SUGGESTIONS)) {
-                    _state.value = SearchUiState.Results(operation = it as SearchUiState.SearchOperation<SearchResults>)
-                }
-
-                if (!FeatureFlag.isEnabled(Feature.IMPROVED_SEARCH_SUGGESTIONS) && it is SearchUiState.SearchOperation.Loading) {
-                    saveSearchTerm(it.searchTerm)
+            if (FeatureFlag.isEnabled(Feature.IMPROVED_SEARCH_RESULTS)) {
+                searchHandler.improvedSearchResults.collect {
                     showSearchHistory = false
+                    _state.value = SearchUiState.ImprovedResults(operation = it)
+                    if (!FeatureFlag.isEnabled(Feature.IMPROVED_SEARCH_SUGGESTIONS) && it is SearchUiState.SearchOperation.Loading) {
+                        saveSearchTerm(it.searchTerm)
+                    }
+                }
+            } else {
+                searchHandler.searchResults.collect {
+                    showSearchHistory = false
+                    if (_state.value is SearchUiState.OldResults) {
+                        _state.value = SearchUiState.OldResults(operation = it)
+                    }
+                    if (!FeatureFlag.isEnabled(Feature.IMPROVED_SEARCH_SUGGESTIONS) && it is SearchUiState.SearchOperation.Loading) {
+                        saveSearchTerm(it.searchTerm)
+                    }
                 }
             }
         }
@@ -97,10 +99,10 @@ class SearchViewModel @Inject constructor(
         // Prevent updating the search query when navigating back to the search results after tapping on a result.
         if (query == _state.value.searchTerm) return
 
-        if (FeatureFlag.isEnabled(Feature.IMPROVED_SEARCH_SUGGESTIONS) && source == SourceView.DISCOVER) {
+        if (FeatureFlag.isEnabled(Feature.IMPROVED_SEARCH_SUGGESTIONS)) {
             searchHandler.updateAutCompleteQuery(query)
             _state.update {
-                if (it is SearchUiState.Results && it.searchTerm.orEmpty().length > query.length) {
+                if ((it is SearchUiState.OldResults || it is SearchUiState.ImprovedResults) && it.searchTerm.orEmpty().length > query.length) {
                     SearchUiState.Suggestions(operation = SearchUiState.SearchOperation.Success(searchTerm = query, results = emptyList()))
                 } else {
                     it
@@ -131,29 +133,32 @@ class SearchViewModel @Inject constructor(
         onSubscribeToPodcast(podcast.uuid)
     }
 
-    suspend fun fetchEpisode(episodeItem: EpisodeItem): BaseEpisode? {
-        return podcastManager.findOrDownloadPodcastRxSingle(episodeItem.podcastUuid)
-            .flatMapMaybe {
-                rxMaybe {
-                    episodeManager.findByUuid(episodeItem.uuid)
-                }
-            }
-            .toFlowable().asFlow().firstOrNull()
-    }
-
     fun selectFilter(filter: ResultsFilters) {
-        if (FeatureFlag.isEnabled(Feature.IMPROVED_SEARCH_RESULTS) && _state.value is SearchUiState.Results) {
+        if (FeatureFlag.isEnabled(Feature.IMPROVED_SEARCH_RESULTS) && _state.value is SearchUiState.ImprovedResults) {
             analyticsTracker.track(
                 AnalyticsEvent.IMPROVED_SEARCH_FILTER_TAPPED,
                 mapOf(
-                    "source" to "discover",
+                    "source" to source,
                     "filter" to filter.name,
                 ),
             )
-            _state.update {
-                (it as SearchUiState.Results).copy(
-                    selectedFilterIndex = ResultsFilters.entries.indexOf(filter),
-                )
+            _state.update { state ->
+                when (state) {
+                    is SearchUiState.ImprovedResults -> {
+                        if (state.operation is SearchUiState.SearchOperation.Success) {
+                            state.copy(
+                                selectedFilterIndex = ResultsFilters.entries.indexOf(filter),
+                                operation = state.operation.copy(
+                                    results = state.operation.results.copy(filter = filter),
+                                ),
+                            )
+                        } else {
+                            state
+                        }
+                    }
+
+                    else -> state
+                }
             }
         }
     }
@@ -164,19 +169,19 @@ class SearchViewModel @Inject constructor(
         // Optimistically update subscribe status
         _state.update {
             when (it) {
-                is SearchUiState.Results -> it.copy(
-                    operation = (it.operation as? SearchUiState.SearchOperation.Success)?.copy(
-                        results = it.operation.results.copy(
-                            podcasts = it.operation.results.podcasts.map { folderItem ->
-                                if (folderItem.uuid == uuid) {
-                                    (folderItem as? FolderItem.Podcast)?.copy(podcast = folderItem.podcast.copy(isSubscribed = true)) ?: folderItem
-                                } else {
-                                    folderItem
-                                }
-                            },
-                        ),
-                    ) ?: it.operation,
-                )
+                is SearchUiState.OldResults ->
+                    it.copy(
+                        operation = (it.operation as? SearchUiState.SearchOperation.Success)?.copy(
+                            results = it.operation.results.subscribeToPodcast(uuid),
+                        ) ?: it.operation,
+                    )
+
+                is SearchUiState.ImprovedResults ->
+                    it.copy(
+                        operation = (it.operation as? SearchUiState.SearchOperation.Success)?.copy(
+                            results = it.operation.results.subscribeToPodcast(uuid),
+                        ) ?: it.operation,
+                    )
 
                 is SearchUiState.Suggestions -> it.copy(
                     operation = (it.operation as? SearchUiState.SearchOperation.Success)?.copy(
@@ -200,19 +205,27 @@ class SearchViewModel @Inject constructor(
         )
     }
 
-    fun selectSuggestion(suggestion: String) {
-        saveSearchTerm(suggestion)
-        searchHandler.updateSearchQuery(suggestion, true)
+    fun runSearchOnTerm(term: String) {
+        saveSearchTerm(term)
+        searchHandler.updateSearchQuery(term, true)
 
+        _state.value = if (FeatureFlag.isEnabled(Feature.IMPROVED_SEARCH_RESULTS)) {
+            SearchUiState.ImprovedResults(operation = SearchUiState.SearchOperation.Loading(term))
+        } else {
+            SearchUiState.OldResults(operation = SearchUiState.SearchOperation.Loading(term))
+        }
+    }
+
+    fun selectSuggestion(suggestion: String) {
         analyticsTracker.track(
             AnalyticsEvent.IMPROVED_SEARCH_SUGGESTION_TERM_TAPPED,
             properties = mapOf(
                 "term" to suggestion,
-                "source" to "discover",
+                "source" to source,
             ),
         )
 
-        _state.value = SearchUiState.Results(operation = SearchUiState.SearchOperation.Loading(suggestion))
+        runSearchOnTerm(suggestion)
     }
 
     fun onFragmentPause(isChangingConfigurations: Boolean?) {
@@ -251,7 +264,7 @@ class SearchViewModel @Inject constructor(
         analyticsTracker.track(
             AnalyticsEvent.IMPROVED_SEARCH_SUGGESTIONS_SHOWN,
             mapOf(
-                "source" to "discover",
+                "source" to source,
             ),
         )
     }
@@ -279,11 +292,45 @@ class SearchViewModel @Inject constructor(
     }
 }
 
-data class SearchResults(
-    val podcasts: List<FolderItem>,
-    val episodes: List<EpisodeItem>,
-) {
-    val isEmpty: Boolean get() = podcasts.isEmpty() && episodes.isEmpty()
+sealed interface SearchResults {
+    val isEmpty: Boolean
+
+    fun subscribeToPodcast(uuid: String): SearchResults
+
+    data class SegregatedResults(
+        val podcasts: List<FolderItem>,
+        val episodes: List<EpisodeItem>,
+    ) : SearchResults {
+        override val isEmpty: Boolean get() = podcasts.isEmpty() && episodes.isEmpty()
+
+        override fun subscribeToPodcast(uuid: String) = copy(
+            podcasts = podcasts.map { folderItem ->
+                if (folderItem.uuid == uuid) {
+                    (folderItem as? FolderItem.Podcast)?.copy(podcast = folderItem.podcast.copy(isSubscribed = true)) ?: folderItem
+                } else {
+                    folderItem
+                }
+            },
+        )
+    }
+
+    data class ImprovedResults(
+        val results: List<ImprovedSearchResultItem>,
+        val filter: ResultsFilters,
+    ) : SearchResults {
+        val filteredResults: List<ImprovedSearchResultItem>
+            get() = results.filter { item ->
+                when (filter) {
+                    ResultsFilters.TOP_RESULTS -> true
+                    ResultsFilters.EPISODES -> item is ImprovedSearchResultItem.EpisodeItem
+                    ResultsFilters.PODCASTS -> item is ImprovedSearchResultItem.PodcastItem || item is ImprovedSearchResultItem.FolderItem
+                }
+            }
+
+        override val isEmpty: Boolean get() = filteredResults.isEmpty()
+
+        override fun subscribeToPodcast(uuid: String) = copy(results = results.map { if (it is ImprovedSearchResultItem.PodcastItem && it.uuid == uuid) it.copy(isFollowed = true) else it })
+    }
 }
 
 enum class ResultsFilters(val resId: Int) {
@@ -297,14 +344,16 @@ sealed interface SearchUiState {
     val searchTerm: String?
         get() = when (this) {
             is Suggestions -> operation.searchTerm
-            is Results -> operation.searchTerm
+            is OldResults -> operation.searchTerm
+            is ImprovedResults -> operation.searchTerm
             else -> null
         }
 
     val isLoading: Boolean
         get() = when (this) {
             is Suggestions -> operation is SearchOperation.Loading
-            is Results -> operation is SearchOperation.Loading
+            is OldResults -> operation is SearchOperation.Loading
+            is ImprovedResults -> operation is SearchOperation.Loading
             else -> false
         }
 
@@ -318,9 +367,13 @@ sealed interface SearchUiState {
 
     data object Idle : SearchUiState
     data class Suggestions(val operation: SearchOperation<List<SearchAutoCompleteItem>>) : SearchUiState
-    data class Results(
-        val operation: SearchOperation<SearchResults>,
+    data class ImprovedResults(
+        val operation: SearchOperation<SearchResults.ImprovedResults>,
         val filterOptions: Set<ResultsFilters> = ResultsFilters.entries.toSet(),
         val selectedFilterIndex: Int = 0,
+    ) : SearchUiState
+
+    data class OldResults(
+        val operation: SearchOperation<SearchResults.SegregatedResults>,
     ) : SearchUiState
 }