+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 *   New Features
     *   Add an option to sort Up Next by date
         ([#4586](https://github.com/Automattic/pocket-casts-android/pull/4586))
+    *   Implement search suggestions
+        ([#4571](https://github.com/Automattic/pocket-casts-android/pull/4571))
 *   Updates
     *   Improve sleep timer terminology and chapter option visibility
         ([#4564](https://github.com/Automattic/pocket-casts-android/pull/4564))